20180831 add security (#10)
Added security using Jwt on server side
  - Configured middleware
  - Created classes to issue tokens
  - Created BaseController to have available information about logged in user
  - Added [Authorize] where needed

Added security on client side
  - Created Login component
  - Created interceptor to include Bearer token on requests
  - Created AuthGuard to check if routes are available
  - Created Observable to notify components about login / logout
  - Modified components to use user from logged in

 - Refactored Repositories
  - Included explicit FK
  - Changed from ChangeTracker to Context.Update

+++ BlogExperimentalPlatform.Web/ClientApp/src/app/blogs/blog-list.component.ts
@@ -1,11 +1,12 @@
-import { Component }            from '@angular/core';
-import { Router }               from "@angular/router";
-import { BaseComponent }        from "../classes/BaseComponent";
-import { NotificationService }  from "../services/notification.service";
+import { Component }              from "@angular/core";
+import { Router }                 from "@angular/router";
+import { BaseComponent }          from "../classes/BaseComponent";
+import { NotificationService }    from "../services/notification.service";
 
-import { BlogDataService }      from './blog-data.service';
+import { BlogDataService }        from "./blog-data.service";
 
-import { IBlog }                from '../Interfaces/IBlog';
+import { IBlog }                  from "../Interfaces/IBlog";
+import { AuthenticationService }  from "../services/authentication.service";
 
 @Component({
   selector: 'blog-list',
@@ -16,10 +17,12 @@ export class BlogListComponent extends BaseComponent {
 
   constructor(
     notificationService: NotificationService,
+    authenticationService: AuthenticationService, 
     private dataService: BlogDataService,
+    
     private router: Router, ) {
 
-    super(notificationService);
+    super(notificationService, authenticationService);
   } 
 
   ngOnInit(): void {
@@ -42,16 +45,22 @@ export class BlogListComponent extends BaseComponent {
   }
 
   removeBlog(blog: IBlog): void {
-    this.notificationService.openConfirmationDialog("Are you sure you want to delete the blog?",
-      () => {
-        this.dataService.deleteBlog(blog.id)
-          .subscribe(() => {
-            this.notificationService.printSuccessMessage(blog.name + " has been deleted.");
-            this.loadBlogs();
-          },
-            (error: any): void => {
-              this.checkForErrorsOnRequest(error, "There's been an error while deleting blog '" + blog.name + "'");
-            });
-      });
+    var user = this.getCurrentUser();
+    if (user && user.id == blog.owner.id) {
+      this.notificationService.openConfirmationDialog("Are you sure you want to delete the blog?",
+        () => {
+          this.dataService.deleteBlog(blog.id)
+            .subscribe(() => {
+              this.notificationService.printSuccessMessage(blog.name + " has been deleted.");
+              this.loadBlogs();
+            },
+              (error: any): void => {
+                this.checkForErrorsOnRequest(error, "There's been an error while deleting blog '" + blog.name + "'");
+              });
+        });
+    }
+    else {
+      this.notificationService.printErrorMessage("You have to be logged in and the owner in order to delete a blog");
+    }
   }
 }
